+++ windows/runner/flutter_window.h
@@ -14,6 +14,8 @@
 // A window that does nothing but host a Flutter view.
 class FlutterWindow : public Win32Window {
  public:
+  static constexpr UINT kRestoreManagementMessage = WM_APP + 17;
+
   // Creates a new FlutterWindow hosting a Flutter view running |project|.
   explicit FlutterWindow(const flutter::DartProject& project);
   virtual ~FlutterWindow();